food app/backend/requirements.txt
flask
torch
ultralytics
pillow
pandas
